backend/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.2